Move Main() into Main.kt.
